requirements.txt
pyTelegramBotAPI
google-generativeai
pandas
psycopg-binary